Add plugin provider context with a cleaner updater
